--- lab1/src/main/java/agh/ics/oop/model/MoveDirection.java
@@ -1,8 +1,0 @@
-package agh.ics.oop.model;
-
-public enum MoveDirection {
-    FORWARD,
-    BACKWARD,
-    RIGHT,
-    LEFT
-}
