+++ oolab/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Oct 22 14:58:48 CEST 2023
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.2-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
