+++ oolab/src/main/java/agh/ics/oop/model/MoveDirection.java
@@ -1,0 +1,8 @@
+package agh.ics.oop.model;
+
+public enum MoveDirection {
+    FORWARD,
+    BACKWARD,
+    RIGHT,
+    LEFT
+}
